[fix] 폴더명 join -> signUp 으로 변경
