+++ web/client/signUp/signUp.html
@@ -5,7 +5,7 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
     
-    <title>join.html</title>
+    <title>signUp.html</title>
     <!--라이브러리-->
     <link rel="stylesheet" href="../boot_css/bootstrap.min.css">
     <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css">
@@ -13,10 +13,14 @@
   
     <!--css 연결-->
     <link rel="stylesheet" href="../boot_css/login.css">
-    <link rel="stylesheet" href="join.css"></style>
+    <link rel="stylesheet" href="signUp.css"></style>
 
     <!--javascript 파일 연결-->
-    <script src="join.js"></script>
+    <script src="signUp.js"></script>
+    <script src="https://cdn.jsdelivr.net/npm/axios/dist/axios.min.js"></script>
+    <script src="./node_modules/axios/dist/axios.min.js"></script>
+    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
+
   </head>
 
   <body>
@@ -25,13 +29,24 @@
             <div class="col xs-12">
             <form method="post" action="./community.html">
                 
-                <!--id-->
+                <!--이름-->
                 <div class="form-group row">
-                    <label class="col-lg-2 col-form-label form-control-label text-muted">ID</label>
+                    <label class="col-lg-2 col-form-label form-control-label text-muted">Name</label>
                         <div class="col-lg-10">
                             <label> <p class="label-txt"></p> 
-                            <input type="text" id="ID" autofocus placeholder="아이디를 입력하세요" class="input" required>
+                            <input type="text" id="NAME" autofocus placeholder="이름을 입력하세요" class="input" required>
+                            <div class="line-box"> <div class="line"></div></div>   
+                    </label></div>
+                </div>
+
+                <!--닉네임-->
+                <div class="form-group row">
+                    <label class="col-lg-2 col-form-label form-control-label text-muted">NickName</label>
+                        <div class="col-lg-10">
+                            <label> <p class="label-txt"></p> 
+                            <input type="text" id="NICKNAME" placeholder="닉네임(ID)을 설정하세요" class="input" > 
                             <div class="line-box"> <div class="line"></div></div> 
+                            <button class="btn btn-outline-primary btn-lg" type="button" id="checkbtn">중복확인</button>     
                     </label> </div>
                 </div>
                 
@@ -40,7 +55,7 @@
                     <label class="col-lg-2 col-form-label form-control-label text-muted">PW</label>
                         <div class="col-lg-10">
                             <label> <p class="label-txt"></p> 
-                            <input type="password" id="PW" placeholder="비밀번호를 입력하세요" class="input" onfocus="goodpw();"required> 
+                            <input type="password" id="PW" placeholder="비밀번호를 입력하세요" class="input" onfocus="goodpw();" required> 
                             <div class="line-box"> <div class="line"></div></div>
                              <!--비밀번호 조건-->
                             <div id="badpw">비밀번호 강도:</div><div id="badpwoption">낮음</div>
@@ -72,20 +87,12 @@
                     </label> </div>
                 </div>
 
-                 <!--닉네임-->
-                 <div class="form-group row">
-                    <label class="col-lg-2 col-form-label form-control-label text-muted">NickName</label>
-                        <div class="col-lg-10">
-                            <label> <p class="label-txt"></p> 
-                            <input type="text" id="nickname" placeholder="닉네임을 설정하세요" class="input" required> 
-                            <div class="line-box"> <div class="line"></div></div> 
-                    </label> </div>
-                </div>
                 <button type="button" class="btn btn-outline-primary btn-lg" onclick="check_join()";>회원가입</button>   
             </form>
            </div>
         </div>
       </div>
+    </body>
     <!--javascript-->
     
     <script src="https://code.jquery.com/jquery.js"></script>
